+++ changelog.md
@@ -45,7 +45,7 @@
   - Display: Adjust word wrap break points to have better flow
   - Update electron 19.1.5 to 22.0.0
   - Update mathjs 11.3.3 to 11.4.0
-  - Update better-sqlite3 7.6.2 to 8.0.0
+  - Update better-sqlite3 7.6.2 to 8.0.1
   - Update fs-extra 10.1.0 to 11.1.0
 
 ## 1.0.0-alpha